test: isolate CLI operation failure regression
